requirements.txt
pydantic~=1.9.0
fastapi~=0.73.0
motor~=2.5.1
python-decouple~=3.5
cryptography~=3.4.7
PyJWT~=2.1.0
geopy~=2.2.0
pymongo~=3.12.3